benchmark/cephfs: use dedicated keyring
To avoid using the admin keyring to access cephfs for benchmarking
purposes a dedicated keyring is introduced. The client nodes only
receive a secrets file (containing the user secret instead of the whole
keyring). This works around issues around salts templating engine and
the way the secret was passed to the mount command.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>