osd.py: always find highest partition for filestore
Currently, for filestore devices, if the journal is on the same
device, osd_partition() is hardcoded to return 2.  This means that
is_prepared() will later return False when run against a disk
where the OSD paritition is actually partition 1.  Instead, let's
just have it always return the highest OSD-type partition number.

Question: can we remove the note that says "Do not remove the logic
with the hardcoded values until the evaluation of activate command
is resolved"?

Signed-off-by: Tim Serong <tserong@suse.com>